# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Team 1 E-Commerce Platform API
  version: "1.0"
  description: |
    Every customer and admin request goes through the API Gateway on `http://localhost:8080` (Kubernetes: `http://localhost:30080`).
    The Gateway validates the Keycloak token, injects the caller's identity, and forwards the request with its own
    client token (ADD §7).

    **Getting a token.** Use the password grant on Keycloak for `customer-test` (CUSTOMER) or `admin-test` (ADMIN).
    The passwords are in your local `deployment/docker/.env`, which is never committed.

    ```
    POST http://localhost:8180/realms/ecommerce-platform/protocol/openid-connect/token
    grant_type=password&client_id=user-sign-in&username=customer-test&password=...
    ```

    Click **Authorize** and paste the `access_token`. Reading products and reviews needs no token.

    **Rate limits (FR-13).** Anonymous callers get 100 requests/s per IP (burst 200). Signed-in callers get 20 requests/s per user (burst 40). Over the limit, the Gateway returns `429`.

    **Internal endpoints** (tag `internal`) have no Gateway route and no host port. Only another service can call them, from
    inside the Compose network or the cluster, with that service's client token.
servers:
  - url: http://localhost:8080
    description: API Gateway (Docker Compose)
  - url: http://localhost:30080
    description: API Gateway (kind)
...
security: []
paths:
...
  /api/v1/inventory/low-stock:
    get:
      operationId: listLowStock
      tags: [inventory]
      summary: List low stock
      description: Stock rows whose available quantity is below `threshold`, lowest first.
      security: [{bearerAuth: []}]
      parameters:
        - {name: threshold, in: query, required: false, schema: {type: integer, minimum: 0, default: 10}}
      responses:
        '200':
          description: Stock rows below the threshold
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/Stock'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
  /api/v1/inventory/{productId}:
    parameters:
      - name: productId
        in: path
        required: true
        schema: {type: integer, format: int64, example: 1}
    get:
      operationId: getStock
      tags: [inventory]
      summary: Read stock
      security: [{bearerAuth: []}]
      responses:
        '200':
          description: Stock for the product
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Stock'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
    put:
      operationId: setStock
      tags: [inventory]
      summary: Set available stock
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/StockAdjustment'}
      responses:
        '200':
          description: Updated stock
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Stock'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /api/v1/inventory/check:
    get:
      operationId: checkStock
      tags: [internal]
      summary: "Internal: stock check before an order is placed"
      description: |
        Caller: `order-service` only, on `http://inventory-service:8084` with its client-credentials token.
        Through the Gateway this path returns `403`.
      servers:
        - url: http://inventory-service:8084
          description: inventory-service inside the Compose network
      security: [{bearerAuth: []}]
      parameters:
        - {name: productId, in: query, required: true, schema: {type: integer, format: int64, minimum: 1, example: 1}}
        - {name: quantity, in: query, required: true, schema: {type: integer, minimum: 1, example: 2}}
      responses:
        '200':
          description: Whether that quantity is available
          content:
            application/json:
              schema: {$ref: '#/components/schemas/StockCheck'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak access token (realm `ecommerce-platform`)
...
  responses:
    BadRequest:
      description: Invalid input (validation message in `message`)
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    Unauthorized:
      description: Missing, expired, or invalid token
    Forbidden:
      description: Token is valid but the role or caller is not allowed
    NotFound:
      description: Not found
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
...
  schemas:
    Error:
      type: object
      properties:
        status: {type: integer, example: 400}
        error: {type: string, example: Bad Request}
        message: {type: string, example: rating must be between 1 and 5}
        path: {type: string, example: /api/v1/products/6/reviews}
...
    Stock:
      type: object
      properties:
        productId: {type: integer, format: int64, example: 1}
        available: {type: integer, example: 100}
        reserved: {type: integer, example: 3}
    StockAdjustment:
      type: object
      required: [available]
      properties:
        available: {type: integer, minimum: 0, example: 150}
    StockCheck:
      type: object
      properties:
        available: {type: boolean, example: true}
...